requirements.txt
dj-database-url==0.4.1
Django==1.10.5
gunicorn==19.6.0
psycopg2==2.6.1
whitenoise==2.0.6
requests==2.12.3
numpy==1.11.2
scipy==0.18.1
scikit-learn==0.18.1
Pillow==3.4.2
opencv-python==3.2.0.6
pytesseract==0.1.6